Update following: 1. Rename service folder to services 2. Remove whitespaces
